Separate out code symbol extraction
