+++ sage/data_manager.py
@@ -4,13 +4,10 @@ import logging
 import os
 from abc import abstractmethod
 from functools import cached_property
-from typing import Any, Dict, Generator, List, Tuple
+from typing import Any, Dict, Generator, Tuple
 
 import requests
 from git import GitCommandError, Repo
-from tree_sitter import Node
-
-from sage.chunker import CodeFileChunker
 
 
 class DataManager:
@@ -178,13 +175,12 @@ class GitHubRepoManager(DataManager):
             )
         return True
 
-    def walk(self, get_content: bool = True, get_code_symbols: bool = False) -> Generator[Tuple[Any, Dict], None, None]:
+    def walk(self, get_content: bool = True) -> Generator[Tuple[Any, Dict], None, None]:
         """Walks the local repository path and yields a tuple of (content, metadata) for each file.
         The filepath is relative to the root of the repository (e.g. "org/repo/your/file/path.py").
 
         Args:
             get_content: When set to True, yields (content, metadata) tuples. When set to False, yields metadata only.
-            get_code_symbols: When set to True, includes code symbols in the metadata, under the "code_symbols" key.
         """
         # We will keep appending to these files during the iteration, so we need to clear them first.
         repo_name = self.repo_id.replace("/", "_")
@@ -217,9 +213,6 @@ class GitHubRepoManager(DataManager):
                     "url": self.url_for_file(relative_file_path),
                 }
 
-                if get_code_symbols:
-                    metadata["code_symbols"] = self._get_code_symbols(relative_file_path)
-
                 if not get_content:
                     yield metadata
                     continue
@@ -228,60 +221,20 @@ class GitHubRepoManager(DataManager):
                 if contents:
                     yield contents, metadata
 
-    def _get_code_symbols(self, file_path: str) -> List[Tuple[str, str]]:
-        """Extracts code symbols from a file.
-
-        Code symbols are tuples of the form (class_name, method_name). For classes, method_name is None. For methods
-        that do not belong to a class, class_name is None.
-        """
-        if not CodeFileChunker.is_code_file(file_path):
-            return []
-
-        content = self.read_file(file_path)
-        if not content:
-            return []
-
-        tree = CodeFileChunker.parse_tree(file_path, content)
-        if not tree:
-            return []
-
-        classes_and_methods = []
-        self._extract_classes_and_methods(tree.root_node, classes_and_methods)
-        return classes_and_methods
-
-    @staticmethod
-    def _extract_classes_and_methods(node: Node, acc: List[Tuple[str, str]], parent_class: str = None):
-        """Extracts classes and methods from a tree-sitter node and places them in the `acc` accumulator."""
-        if node.type in ["class_definition", "class_declaration"]:
-            class_name_node = node.child_by_field_name("name")
-            if class_name_node:
-                class_name = class_name_node.text.decode("utf-8")
-                acc.append((class_name, None))
-                for child in node.children:
-                    GitHubRepoManager._extract_classes_and_methods(child, acc, class_name)
-        elif node.type in ["function_definition", "method_definition"]:
-            function_name_node = node.child_by_field_name("name")
-            if function_name_node:
-                acc.append((parent_class, function_name_node.text.decode("utf-8")))
-                # We're not going deeper into a method. This means we're missing nested functions.
-        else:
-            for child in node.children:
-                GitHubRepoManager._extract_classes_and_methods(child, acc, parent_class)
-
     def url_for_file(self, file_path: str) -> str:
         """Converts a repository file path to a GitHub link."""
         file_path = file_path[len(self.repo_id) + 1 :]
         return f"https://github.com/{self.repo_id}/blob/{self.default_branch}/{file_path}"
 
     def read_file(self, relative_file_path: str) -> str:
-        """Reads the content of the file at the given path."""
-        file_path = os.path.join(self.local_dir, relative_file_path)
-        with open(file_path, "r") as f:
+        """Reads the contents of a file in the repository."""
+        absolute_file_path = os.path.join(self.local_dir, relative_file_path)
+        with open(absolute_file_path, "r") as f:
             try:
                 contents = f.read()
                 return contents
             except UnicodeDecodeError:
-                logging.warning("Unable to decode file %s.", file_path)
+                logging.warning("Unable to decode file %s.", absolute_file_path)
                 return None
 
     def from_args(args: Dict):
